requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
distlib==0.3.1
Django==3.2
djangorestframework==3.12.4
filelock==3.0.12
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.16.0
pylint==2.6.0
pyodbc==4.0.30
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
virtualenv==20.4.4
wrapt==1.12.1
